Change float to double, small code refactor

+++ Expression String Parser/Source.cpp
@@ -1,4 +1,5 @@
 #include <chrono>
+#include <iostream>
 #include <iomanip>  // << std::setprecision(10)
 
 #include "ExprStrParser.h"
@@ -10,13 +11,13 @@ int main() {
 	//std::string expr = "log(x)";
 	//std::string expr = "cos(x)^2 + log(x/3) * round(4.5*(sin(x)))";
 
-	//std::map<std::string, float> args{{"x", 2.0f} };
+	//std::map<std::string, double> args{{"x", 2.0f} };
 
 	auto start = std::chrono::high_resolution_clock::now();
 	parser.Parse(expr);
 	auto stop = std::chrono::high_resolution_clock::now();
 	auto duration = std::chrono::duration_cast<std::chrono::microseconds>(stop - start);
-	std::cout << "Time taken to parse and set func: "<< static_cast<float>(duration.count())/1000.0f << " milliseconds" << std::endl;
+	std::cout << "Time taken to parse and set func: "<< static_cast<double>(duration.count())/1000.0 << " milliseconds" << std::endl;
 
 	std::cout << "Answer: " << std::fixed << std::setprecision(10) << parser.Calculate(3.0f) << std::defaultfloat << std::endl;
 	constexpr std::size_t calc_count = 100000;
@@ -26,6 +27,6 @@ int main() {
 	}
 	stop = std::chrono::high_resolution_clock::now();
 	duration = std::chrono::duration_cast<std::chrono::microseconds>(stop - start);
-	std::cout << "Time taken to calculate "<<calc_count<< " times: "<< static_cast<float>(duration.count())/1000.0f << " milliseconds" << std::endl;
+	std::cout << "Time taken to calculate "<<calc_count<< " times: "<< static_cast<double>(duration.count())/1000.0 << " milliseconds" << std::endl;
 	return 0;
 }
